+++ shared/domain/src/commonMain/kotlin/app/hdj/datepick/domain/usecase/featured/GetFeaturedUseCase.kt
@@ -9,6 +9,6 @@ class GetFeaturedUseCase @Inject constructor(
     private val featuredRepository: FeaturedRepository
 ) {
 
-    fun execute(id: Int) = featuredRepository.getFeaturedDetail(id)
+    fun execute(id: Long) = featuredRepository.getFeaturedDetail(id)
 
 }
